Optimizing css files in production mode.
