+++ package.json
@@ -7,6 +7,7 @@
 		"@rollup/plugin-terser": "^0.4.4",
 		"@tsconfig/svelte": "^5.0.4",
 		"@types/firefox-webext-browser": "^120.0.4",
+		"esbuild": "^0.25.1",
 		"hex-to-css-filter": "^6.0.0",
 		"prettier": "^3.3.3",
 		"prettier-plugin-svelte": "^3.2.7",
